Shade protobuf and release the assembly jar (#38)
v0.3.0 doesn't work on Spark 2.3 because the latter has protobuf 2.x while CoreNLP 3.9.1 depends on protobuf 3.x.

This PR changes build to create an assembly jar with protobuf shaded. We cannot shade CoreNLP because it seems using reflection. Other transitive dependencies are kept. We might need to improve shading in a follow-up work.

Closes #37